// src/interfaces/briefing-type.ts
export type Briefing = {
    id:                                   number;
    case_id:                              string;
    user_id:                              number;
    user_name:                            string;
    process_id:                           number;
    process_name:                         string;
    process_owner_id:                     number;
    process_owner_name:                   string;
    nombrecliente:                        string;
    nombremarca:                          string;
    personacontacto:                      string;
    telefonocontacto:                     string;
    emailcontacto:                        string;
    negocioempresa:                       string;
    historiaempresa:                      string;
    ubicacionservicio:                    string;
    necesidadquesatisface:                string;
    diferenciador:                        string;
    targetedad:                           string;
    targetgenero:                         string;
    targetlugardeorigen:                  string;
    targetgustos:                         string;
    targetactitudes:                      string;
    targetestilodevida:                   string;
    competidores:                         string;
    marcapersonaedad:                     string;
    marcapersonagenero:                   string;
    marcapersonavestido:                  string;
    marcapersonacarro:                    string;
    detallerequerimientotipo:             string;
    detallerequerimientoexpresaimagen:    string;
    detallerequerimientopreferenciacolor: string;
    detallerequerimientocomopercepcion:   string;
    detallerequerimientoadicional:        string;
    mediosimpresos:                       string;
    rrss:                                 string;
    audiovisual:                          string;
    created_at:                           Date;
}